+++ Content/SkillPassives/RainOfArrowsPassives/FesteringSpores.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
+
+internal class FesteringSpores(SkillTree tree) : SkillPassive(tree);
