+++ Content/SkillPassives/RainOfArrowsPassives/LingeringPoison.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
+
+internal class LingeringPoison(SkillTree tree) : SkillPassive(tree);
