+++ Content/SkillPassives/RainOfArrowsPassives/SharpenedTips.cs
@@ -1,0 +1,9 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
+
+internal class SharpenedTips(SkillTree tree) : SkillPassive(tree)
+{
+	public override object[] TooltipArguments => ["10"];
+}
