+++ Content/SkillPassives/RainOfArrowsPassives/SlicingShrapnel.cs
@@ -1,0 +1,9 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsPassives;
+
+internal class SlicingShrapnel(SkillTree tree) : SkillPassive(tree)
+{
+	public override object[] TooltipArguments => ["3"];
+}
