+++ Content/SkillPassives/SwarmPassives/AggressiveChill.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
+
+internal class AggressiveChill(SkillTree tree) : SkillPassive(tree);
