+++ Content/SkillPassives/SwarmPassives/Eggsplosion.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
+
+internal class Eggsplosion(SkillTree tree) : SkillPassive(tree);
