+++ Content/SkillPassives/SwarmPassives/IceVenom.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
+
+internal class IceVenom(SkillTree tree) : SkillPassive(tree);
