+++ Content/SkillPassives/SwarmPassives/ShatteringCarapace.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
+
+internal class ShatteringCarapace(SkillTree tree) : SkillPassive(tree);
