+++ Content/SkillPassives/SwarmPassives/ViciousBites.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
+
+internal class ViciousBites(SkillTree tree) : SkillPassive(tree);
