+++ Content/SkillPassives/SwarmPassives/VolatileInsects.cs
@@ -1,0 +1,9 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
+
+internal class VolatileInsects(SkillTree tree) : SkillPassive(tree)
+{
+	public override object[] TooltipArguments => ["300"];
+}
